# specs/001-implement-todo-backend/contracts/openapi.yml
openapi: 3.0.0
info:
  title: Todo App API
  description: API for managing tasks in the Todo App.
  version: 1.0.0

paths:
  /api/tasks:
    get:
      summary: List all tasks for the authenticated user
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [all, pending, completed]
          required: false
      responses:
        '200':
          description: A list of tasks.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TaskRead'
        '401':
          description: Unauthorized
    post:
      summary: Create a new task
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreate'
      responses:
        '200':
          description: The created task.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskRead'
        '400':
          description: Validation Error
        '401':
          description: Unauthorized

  /api/tasks/{id}:
    get:
      summary: Get a specific task by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: The requested task.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskRead'
        '401':
          description: Unauthorized
        '404':
          description: Not Found
    put:
      summary: Update a task
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdate'
      responses:
        '200':
          description: The updated task.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskRead'
        '401':
          description: Unauthorized
        '404':
          description: Not Found
    delete:
      summary: Delete a task
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Task deleted successfully.
        '401':
          description: Unauthorized
        '404':
          description: Not Found

  /api/tasks/{id}/complete:
    patch:
      summary: Toggle a task's completion status
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: The updated task.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskRead'
        '401':
          description: Unauthorized
        '404':
          description: Not Found

components:
  schemas:
    TaskBase:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: string
          maxLength: 1000
        completed:
          type: boolean
          default: false
    TaskCreate:
      allOf:
        - $ref: '#/components/schemas/TaskBase'
        - type: object
          required:
            - title
    TaskUpdate:
      allOf:
        - $ref: '#/components/schemas/TaskBase'
    TaskRead:
      allOf:
        - $ref: '#/components/schemas/TaskBase'
        - type: object
          properties:
            id:
              type: integer
            created_at:
              type: string
              format: date-time
            updated_at:
              type: string
              format: date-time
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT